+++ cmonge/datasets/single_loader.py
@@ -13,7 +13,6 @@ from jaxtyping import PRNGKeyArray
 from loguru import logger
 from sklearn.model_selection import train_test_split
 
-from cmonge.datasets.chemCPA_loaders import SciPlexCPAModule
 from cmonge.trainers.ae_trainer import AETrainerModule
 from cmonge.utils import load_config
 
@@ -408,10 +407,3 @@ class FourIModule(AbstractDataModule):
         else:
             self.encoder = lambda x: x
             self.decoder = lambda x: x
-
-
-DataModuleFactory = {
-    "4i": FourIModule,
-    "sciplex": SciPlexModule,
-    "cpa_sciplex": SciPlexCPAModule,
-}  # , "synthetic": SyntheticDosageModule}
